requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via pytest
brine==1.0.0
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
dataclasses==0.6
decorator==4.3.2          # via networkx
future==0.17.1
graphviz==0.8.4           # via mxnet-cu100
idna==2.8                 # via requests
kiwisolver==1.0.1         # via matplotlib
matplotlib==3.0.2
more-itertools==5.0.0     # via pytest
mxnet-cu100==1.3.1
mypy-extensions==0.4.1
mypy==0.660
networkx==2.2
numpy==1.14.6
pluggy==0.8.1             # via pytest
py==1.7.0                 # via pytest
pyparsing==2.3.1          # via matplotlib
pytest==4.2.0
python-dateutil==2.7.5    # via matplotlib
requests==2.21.0          # via mxnet-cu100
six==1.12.0
tqdm==4.30.0
typed-ast==1.2.0          # via mypy
urllib3==1.24.1           # via requests
